+++ src/main/java/ru/java/mentor/oldranger/club/service/forum/CommentService.java
@@ -1,0 +1,5 @@
+package ru.java.mentor.oldranger.club.service.forum;
+
+public interface CommentService {
+
+}
