+++ src/main/java/ru/java/mentor/oldranger/club/service/forum/SectionService.java
@@ -1,0 +1,8 @@
+package ru.java.mentor.oldranger.club.service.forum;
+
+import ru.java.mentor.oldranger.club.model.forum.Section;
+
+public interface SectionService {
+
+    public void addSection(Section section);
+}
