+++ src/main/java/ru/java/mentor/oldranger/club/service/forum/impl/CommentServiceImpl.java
@@ -1,0 +1,13 @@
+package ru.java.mentor.oldranger.club.service.forum.impl;
+
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Service;
+import ru.java.mentor.oldranger.club.dao.ForumRepository.CommentRepository;
+import ru.java.mentor.oldranger.club.service.forum.CommentService;
+
+@Service
+public class CommentServiceImpl implements CommentService {
+
+    @Autowired
+    private CommentRepository commentRepository;
+}
