+++ src/main/java/ru/java/mentor/oldranger/club/service/forum/impl/TopicServiceImpl.java
@@ -1,0 +1,29 @@
+package ru.java.mentor.oldranger.club.service.forum.impl;
+
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Service;
+import ru.java.mentor.oldranger.club.dao.ForumRepository.TopicRepository;
+import ru.java.mentor.oldranger.club.model.forum.Topic;
+import ru.java.mentor.oldranger.club.service.forum.TopicService;
+
+@Service
+public class TopicServiceImpl implements TopicService {
+
+    @Autowired
+    private TopicRepository topicRepository;
+
+    @Override
+    public void createTopic(Topic topic) {
+        topicRepository.save(topic);
+    }
+
+    @Override
+    public void editTopicByName(Topic topic) {
+        topicRepository.save(topic);
+    }
+
+    @Override
+    public void deleteTopicById(Long id) {
+        topicRepository.deleteById(id);
+    }
+}
