+++ src/main/java/ru/java/mentor/oldranger/club/service/mail/MailService.java
@@ -1,0 +1,5 @@
+package ru.java.mentor.oldranger.club.service.mail;
+
+public interface MailService {
+    void send(String to, String subject, String message);
+}
