+++ src/main/java/ru/java/mentor/oldranger/club/service/news/NewsService.java
@@ -1,0 +1,10 @@
+package ru.java.mentor.oldranger.club.service.news;
+
+import ru.java.mentor.oldranger.club.model.News;
+
+import java.util.List;
+
+public interface NewsService {
+
+    public List<News> getAllNews();
+}
