+++ src/main/java/ru/java/mentor/oldranger/club/service/user/RoleService.java
@@ -1,0 +1,5 @@
+package ru.java.mentor.oldranger.club.service.user;
+
+public interface RoleService {
+
+}
