+++ src/main/java/ru/java/mentor/oldranger/club/service/user/UserProfileService.java
@@ -1,0 +1,12 @@
+package ru.java.mentor.oldranger.club.service.user;
+
+import ru.java.mentor.oldranger.club.model.user.UserProfile;
+
+public interface UserProfileService {
+
+    public UserProfile createUserProfile(UserProfile userProfile);
+
+    public void deleteUserProfileById(Long id);
+
+    public void editUserProfile(UserProfile userProfile);
+}
