+++ src/main/java/ru/java/mentor/oldranger/club/service/user/UserStaticService.java
@@ -1,0 +1,10 @@
+package ru.java.mentor.oldranger.club.service.user;
+
+import ru.java.mentor.oldranger.club.model.user.UserStatistic;
+
+public interface UserStaticService {
+
+    public UserStatistic getUserStaticById(Long id);
+
+    public void saveUserStatic(UserStatistic userStatistic);
+}
