+++ src/main/java/ru/java/mentor/oldranger/club/service/user/impl/RoleServiceImpl.java
@@ -1,0 +1,13 @@
+package ru.java.mentor.oldranger.club.service.user.impl;
+
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Service;
+import ru.java.mentor.oldranger.club.dao.UserRepository.RoleRepository;
+import ru.java.mentor.oldranger.club.service.user.RoleService;
+
+@Service
+public class RoleServiceImpl implements RoleService {
+
+    @Autowired
+    private RoleRepository roleRepository;
+}
