--- src/main/java/ru/java/mentor/oldranger/club/sevice/ChatService/MessageService.java
@@ -1,11 +1,0 @@
-package ru.java.mentor.oldranger.club.sevice.ChatService;
-
-import ru.java.mentor.oldranger.club.model.chat.Message;
-
-public interface MessageService {
-
-    public void addMessage(Message message);
-
-    public void removeMessageById(Long id);
-
-}
