--- src/main/java/ru/java/mentor/oldranger/club/sevice/ForumService/CommentService.java
@@ -1,5 +1,0 @@
-package ru.java.mentor.oldranger.club.sevice.ForumService;
-
-public interface CommentService {
-
-}
