--- src/main/java/ru/java/mentor/oldranger/club/sevice/ForumService/Impl/CommentServiceImpl.java
@@ -1,13 +1,0 @@
-package ru.java.mentor.oldranger.club.sevice.ForumService.Impl;
-
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.stereotype.Service;
-import ru.java.mentor.oldranger.club.dao.ForumRepository.CommentRepository;
-import ru.java.mentor.oldranger.club.sevice.ForumService.CommentService;
-
-@Service
-public class CommentServiceImpl implements CommentService {
-
-    @Autowired
-    private CommentRepository commentRepository;
-}
