--- src/main/java/ru/java/mentor/oldranger/club/sevice/ForumService/SectionService.java
@@ -1,8 +1,0 @@
-package ru.java.mentor.oldranger.club.sevice.ForumService;
-
-import ru.java.mentor.oldranger.club.model.forum.Section;
-
-public interface SectionService {
-
-    public void addSection(Section section);
-}
