--- src/main/java/ru/java/mentor/oldranger/club/sevice/MailService/Impl/MailServiceImpl.java
@@ -1,32 +1,0 @@
-package ru.java.mentor.oldranger.club.sevice.MailService.Impl;
-
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.mail.SimpleMailMessage;
-import org.springframework.mail.javamail.JavaMailSender;
-import org.springframework.stereotype.Service;
-import ru.java.mentor.oldranger.club.sevice.MailService.MailService;
-
-@Service
-public class MailServiceImpl implements MailService {
-
-    @Autowired
-    private JavaMailSender mailSender;
-
-    @Value("${spring.mail.username}")
-    private String username;
-
-    @Override
-    public void send(String to, String subject, String message) {
-        SimpleMailMessage mailMessage = new SimpleMailMessage();
-
-        mailMessage.setFrom(username);
-        mailMessage.setTo(to);
-        mailMessage.setSubject(subject);
-        mailMessage.setText(message);
-
-        mailSender.send(mailMessage);
-    }
-}
-
-
