--- src/main/java/ru/java/mentor/oldranger/club/sevice/MailService/MailService.java
@@ -1,5 +1,0 @@
-package ru.java.mentor.oldranger.club.sevice.MailService;
-
-public interface MailService {
-    void send(String to, String subject, String message);
-}
