--- src/main/java/ru/java/mentor/oldranger/club/sevice/NewsService/NewsService.java
@@ -1,10 +1,0 @@
-package ru.java.mentor.oldranger.club.sevice.NewsService;
-
-import ru.java.mentor.oldranger.club.model.News;
-
-import java.util.List;
-
-public interface NewsService {
-
-    public List<News> getAllNews();
-}
