--- src/main/java/ru/java/mentor/oldranger/club/sevice/UserService/RoleService.java
@@ -1,5 +1,0 @@
-package ru.java.mentor.oldranger.club.sevice.UserService;
-
-public interface RoleService {
-
-}
