--- src/main/java/ru/java/mentor/oldranger/club/sevice/UserService/UserStaticService.java
@@ -1,10 +1,0 @@
-package ru.java.mentor.oldranger.club.sevice.UserService;
-
-import ru.java.mentor.oldranger.club.model.user.UserStatistic;
-
-public interface UserStaticService {
-
-    public UserStatistic getUserStaticById(Long id);
-
-    public void saveUserStatic(UserStatistic userStatistic);
-}
